LayoutTests/fast/regions/auto-height/first-region-auto-height-no-breaks.html

<!doctype html>
<html>
    <head>
        <!--
        Test for regions auto-height prototype. All p's are flowed into the same first region, which
        has auto-height and eats them up. No p's should go into the second region.
        -->
        <script src="../../../fast/js/resources/js-test-pre.js"></script>
        <style>
            body { font-family: monospace; }
            #region, #region2 { -webkit-flow-from: flow; border: 1px solid red; }
            #p1, #p2, #p3, #p4 { -webkit-flow-into: flow; }
            #region    { width: 80px; }
            #region2    { width: 120px; height: 50px; }
            p { border: 1px solid green; }
        </style>
    </head>
    <body>
        <p id="p1">Text inside p1.</p>
        <p id="p2">Text inside p2.</p>
        <p id="p3">Text inside p3.</p>
        <p id="p4">Text inside p4.</p>
        <div id="region"></div>
        <div id="region2"></div>
        <script>
            if (window.layoutTestController)
                window.layoutTestController.dumpAsText();
            description("Test for regions auto height prototype.");

            var namedFlow = document.webkitGetFlowByName("flow");
            var p1RegionList = namedFlow.getRegionsByContentNode(document.getElementById("p1"));
            var p2RegionList = namedFlow.getRegionsByContentNode(document.getElementById("p2"));
            var p3RegionList = namedFlow.getRegionsByContentNode(document.getElementById("p3"));
            var p4RegionList = namedFlow.getRegionsByContentNode(document.getElementById("p4"));
            shouldEvaluateTo("p1RegionList.length", 1);
            shouldBeEqualToString("p1RegionList.item(0).id", "region");
            shouldEvaluateTo("p2RegionList.length", 1);
            shouldBeEqualToString("p2RegionList.item(0).id", "region"); 
            shouldEvaluateTo("p3RegionList.length", 1);
            shouldBeEqualToString("p3RegionList.item(0).id", "region");
            shouldEvaluateTo("p4RegionList.length", 1);
            shouldBeEqualToString("p4RegionList.item(0).id", "region"); 


            document.getElementById("p1").style.visibility = "hidden";
            document.getElementById("p2").style.visibility = "hidden";
            document.getElementById("p3").style.visibility = "hidden";
            document.getElementById("p4").style.visibility = "hidden";

            document.getElementById("region").style.visibility = "hidden";
            document.getElementById("region2").style.visibility = "hidden";
        </script>
    </body>
    <script src="../../../fast/js/resources/js-test-post.js"></script>
</html>
